+++ mall-mbg/src/main/java/com/macro/mall/mapper/CmsMemberReportMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsMemberReportMapper {
-    long countByExample(CmsMemberReportExample example);
+    int countByExample(CmsMemberReportExample example);
 
     int deleteByExample(CmsMemberReportExample example);
