+++ mall-mbg/src/main/java/com/macro/mall/mapper/CmsPrefrenceAreaMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsPrefrenceAreaMapper {
-    long countByExample(CmsPrefrenceAreaExample example);
+    int countByExample(CmsPrefrenceAreaExample example);
 
     int deleteByExample(CmsPrefrenceAreaExample example);
