+++ mall-mbg/src/main/java/com/macro/mall/mapper/CmsSubjectCommentMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsSubjectCommentMapper {
-    long countByExample(CmsSubjectCommentExample example);
+    int countByExample(CmsSubjectCommentExample example);
 
     int deleteByExample(CmsSubjectCommentExample example);
