+++ mall-mbg/src/main/java/com/macro/mall/mapper/CmsSubjectMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsSubjectMapper {
-    long countByExample(CmsSubjectExample example);
+    int countByExample(CmsSubjectExample example);
 
     int deleteByExample(CmsSubjectExample example);
