+++ mall-mbg/src/main/java/com/macro/mall/mapper/CmsTopicMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CmsTopicMapper {
-    long countByExample(CmsTopicExample example);
+    int countByExample(CmsTopicExample example);
 
     int deleteByExample(CmsTopicExample example);
