+++ mall-mbg/src/main/java/com/macro/mall/mapper/OmsCartItemMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OmsCartItemMapper {
-    long countByExample(OmsCartItemExample example);
+    int countByExample(OmsCartItemExample example);
 
     int deleteByExample(OmsCartItemExample example);
