+++ mall-mbg/src/main/java/com/macro/mall/mapper/OmsCompanyAddressMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OmsCompanyAddressMapper {
-    long countByExample(OmsCompanyAddressExample example);
+    int countByExample(OmsCompanyAddressExample example);
 
     int deleteByExample(OmsCompanyAddressExample example);
