+++ mall-mbg/src/main/java/com/macro/mall/mapper/OmsOrderMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OmsOrderMapper {
-    long countByExample(OmsOrderExample example);
+    int countByExample(OmsOrderExample example);
 
     int deleteByExample(OmsOrderExample example);
