+++ mall-mbg/src/main/java/com/macro/mall/mapper/PmsBrandMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PmsBrandMapper {
-    long countByExample(PmsBrandExample example);
+    int countByExample(PmsBrandExample example);
 
     int deleteByExample(PmsBrandExample example);
