+++ mall-mbg/src/main/java/com/macro/mall/mapper/PmsProductAttributeCategoryMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PmsProductAttributeCategoryMapper {
-    long countByExample(PmsProductAttributeCategoryExample example);
+    int countByExample(PmsProductAttributeCategoryExample example);
 
     int deleteByExample(PmsProductAttributeCategoryExample example);
