+++ mall-mbg/src/main/java/com/macro/mall/mapper/SmsCouponHistoryMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SmsCouponHistoryMapper {
-    long countByExample(SmsCouponHistoryExample example);
+    int countByExample(SmsCouponHistoryExample example);
 
     int deleteByExample(SmsCouponHistoryExample example);
