+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsAdminLoginLogMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsAdminLoginLogMapper {
-    long countByExample(UmsAdminLoginLogExample example);
+    int countByExample(UmsAdminLoginLogExample example);
 
     int deleteByExample(UmsAdminLoginLogExample example);
