+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsAdminPermissionRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsAdminPermissionRelationMapper {
-    long countByExample(UmsAdminPermissionRelationExample example);
+    int countByExample(UmsAdminPermissionRelationExample example);
 
     int deleteByExample(UmsAdminPermissionRelationExample example);
