+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsAdminRoleRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsAdminRoleRelationMapper {
-    long countByExample(UmsAdminRoleRelationExample example);
+    int countByExample(UmsAdminRoleRelationExample example);
 
     int deleteByExample(UmsAdminRoleRelationExample example);
