+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsMemberProductCategoryRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsMemberProductCategoryRelationMapper {
-    long countByExample(UmsMemberProductCategoryRelationExample example);
+    int countByExample(UmsMemberProductCategoryRelationExample example);
 
     int deleteByExample(UmsMemberProductCategoryRelationExample example);
