+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsMemberReceiveAddressMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsMemberReceiveAddressMapper {
-    long countByExample(UmsMemberReceiveAddressExample example);
+    int countByExample(UmsMemberReceiveAddressExample example);
 
     int deleteByExample(UmsMemberReceiveAddressExample example);
