+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsMemberStatisticsInfoMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsMemberStatisticsInfoMapper {
-    long countByExample(UmsMemberStatisticsInfoExample example);
+    int countByExample(UmsMemberStatisticsInfoExample example);
 
     int deleteByExample(UmsMemberStatisticsInfoExample example);
