+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsPermissionMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsPermissionMapper {
-    long countByExample(UmsPermissionExample example);
+    int countByExample(UmsPermissionExample example);
 
     int deleteByExample(UmsPermissionExample example);
