+++ mall-mbg/src/main/java/com/macro/mall/mapper/UmsRolePermissionRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UmsRolePermissionRelationMapper {
-    long countByExample(UmsRolePermissionRelationExample example);
+    int countByExample(UmsRolePermissionRelationExample example);
 
     int deleteByExample(UmsRolePermissionRelationExample example);
